[API] Test Runner: Refactors wipe cluser into its own file
