+++ elasticsearch-api/spec/rest_api_tests_helper.rb
@@ -17,6 +17,7 @@
 
 require "#{File.expand_path(File.dirname('..'))}/api-spec-testing/test_file"
 require "#{File.expand_path(File.dirname('..'))}/api-spec-testing/rspec_matchers"
+require "#{File.expand_path(File.dirname('..'))}/api-spec-testing/wipe_cluster"
 include Elasticsearch::RestAPIYAMLTests
 
 PROJECT_PATH = File.join(File.dirname(__FILE__), '..')
